+++ packages/types/src/index.ts
@@ -1,3 +1,4 @@
 export * from './disposable'
 export * from './error'
 export * from './path'
+export * from './subscriber'